+++ lib/models/user_model.freezed.dart
@@ -19,7 +19,8 @@ mixin _$UserModel {
  int get currentPoints; String? get homeBaseId; String? get qrToken; String? get bio; String? get photoUrl; String? get bannerUrl; List<String> get following; String get realNameVisibility;// 'Private', 'Friends Only', 'Everyone'
  String get onlineStatus; String? get currentCheckInHallId; List<String> get blockedUsers;// Array of User IDs this user has blocked
  List<String> get customCategories;// Array of Custom Tags/Categories the user has saved
-@NullableTimestampConverter() DateTime? get lastViewedPhotoApprovals;
+@NullableTimestampConverter() DateTime? get lastViewedPhotoApprovals;// Track when the worker last viewed pending photos
+ List<String> get fcmTokens;
 /// Create a copy of UserModel
 /// with the given fields replaced by the non-null parameter values.
 @JsonKey(includeFromJson: false, includeToJson: false)
@@ -32,16 +33,16 @@ $UserModelCopyWith<UserModel> get copyWith => _$UserModelCopyWithImpl<UserModel>
 
 @override
 bool operator ==(Object other) {
-  return identical(this, other) || (other.runtimeType == runtimeType&&other is UserModel&&(identical(other.uid, uid) || other.uid == uid)&&(identical(other.email, email) || other.email == email)&&(identical(other.firstName, firstName) || other.firstName == firstName)&&(identical(other.lastName, lastName) || other.lastName == lastName)&&(identical(other.username, username) || other.username == username)&&(identical(other.birthday, birthday) || other.birthday == birthday)&&(identical(other.phoneNumber, phoneNumber) || other.phoneNumber == phoneNumber)&&(identical(other.recoveryEmail, recoveryEmail) || other.recoveryEmail == recoveryEmail)&&(identical(other.role, role) || other.role == role)&&(identical(other.currentPoints, currentPoints) || other.currentPoints == currentPoints)&&(identical(other.homeBaseId, homeBaseId) || other.homeBaseId == homeBaseId)&&(identical(other.qrToken, qrToken) || other.qrToken == qrToken)&&(identical(other.bio, bio) || other.bio == bio)&&(identical(other.photoUrl, photoUrl) || other.photoUrl == photoUrl)&&(identical(other.bannerUrl, bannerUrl) || other.bannerUrl == bannerUrl)&&const DeepCollectionEquality().equals(other.following, following)&&(identical(other.realNameVisibility, realNameVisibility) || other.realNameVisibility == realNameVisibility)&&(identical(other.onlineStatus, onlineStatus) || other.onlineStatus == onlineStatus)&&(identical(other.currentCheckInHallId, currentCheckInHallId) || other.currentCheckInHallId == currentCheckInHallId)&&const DeepCollectionEquality().equals(other.blockedUsers, blockedUsers)&&const DeepCollectionEquality().equals(other.customCategories, customCategories)&&(identical(other.lastViewedPhotoApprovals, lastViewedPhotoApprovals) || other.lastViewedPhotoApprovals == lastViewedPhotoApprovals));
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is UserModel&&(identical(other.uid, uid) || other.uid == uid)&&(identical(other.email, email) || other.email == email)&&(identical(other.firstName, firstName) || other.firstName == firstName)&&(identical(other.lastName, lastName) || other.lastName == lastName)&&(identical(other.username, username) || other.username == username)&&(identical(other.birthday, birthday) || other.birthday == birthday)&&(identical(other.phoneNumber, phoneNumber) || other.phoneNumber == phoneNumber)&&(identical(other.recoveryEmail, recoveryEmail) || other.recoveryEmail == recoveryEmail)&&(identical(other.role, role) || other.role == role)&&(identical(other.currentPoints, currentPoints) || other.currentPoints == currentPoints)&&(identical(other.homeBaseId, homeBaseId) || other.homeBaseId == homeBaseId)&&(identical(other.qrToken, qrToken) || other.qrToken == qrToken)&&(identical(other.bio, bio) || other.bio == bio)&&(identical(other.photoUrl, photoUrl) || other.photoUrl == photoUrl)&&(identical(other.bannerUrl, bannerUrl) || other.bannerUrl == bannerUrl)&&const DeepCollectionEquality().equals(other.following, following)&&(identical(other.realNameVisibility, realNameVisibility) || other.realNameVisibility == realNameVisibility)&&(identical(other.onlineStatus, onlineStatus) || other.onlineStatus == onlineStatus)&&(identical(other.currentCheckInHallId, currentCheckInHallId) || other.currentCheckInHallId == currentCheckInHallId)&&const DeepCollectionEquality().equals(other.blockedUsers, blockedUsers)&&const DeepCollectionEquality().equals(other.customCategories, customCategories)&&(identical(other.lastViewedPhotoApprovals, lastViewedPhotoApprovals) || other.lastViewedPhotoApprovals == lastViewedPhotoApprovals)&&const DeepCollectionEquality().equals(other.fcmTokens, fcmTokens));
 }
 
 @JsonKey(includeFromJson: false, includeToJson: false)
 @override
-int get hashCode => Object.hashAll([runtimeType,uid,email,firstName,lastName,username,birthday,phoneNumber,recoveryEmail,role,currentPoints,homeBaseId,qrToken,bio,photoUrl,bannerUrl,const DeepCollectionEquality().hash(following),realNameVisibility,onlineStatus,currentCheckInHallId,const DeepCollectionEquality().hash(blockedUsers),const DeepCollectionEquality().hash(customCategories),lastViewedPhotoApprovals]);
+int get hashCode => Object.hashAll([runtimeType,uid,email,firstName,lastName,username,birthday,phoneNumber,recoveryEmail,role,currentPoints,homeBaseId,qrToken,bio,photoUrl,bannerUrl,const DeepCollectionEquality().hash(following),realNameVisibility,onlineStatus,currentCheckInHallId,const DeepCollectionEquality().hash(blockedUsers),const DeepCollectionEquality().hash(customCategories),lastViewedPhotoApprovals,const DeepCollectionEquality().hash(fcmTokens)]);
 
 @override
 String toString() {
-  return 'UserModel(uid: $uid, email: $email, firstName: $firstName, lastName: $lastName, username: $username, birthday: $birthday, phoneNumber: $phoneNumber, recoveryEmail: $recoveryEmail, role: $role, currentPoints: $currentPoints, homeBaseId: $homeBaseId, qrToken: $qrToken, bio: $bio, photoUrl: $photoUrl, bannerUrl: $bannerUrl, following: $following, realNameVisibility: $realNameVisibility, onlineStatus: $onlineStatus, currentCheckInHallId: $currentCheckInHallId, blockedUsers: $blockedUsers, customCategories: $customCategories, lastViewedPhotoApprovals: $lastViewedPhotoApprovals)';
+  return 'UserModel(uid: $uid, email: $email, firstName: $firstName, lastName: $lastName, username: $username, birthday: $birthday, phoneNumber: $phoneNumber, recoveryEmail: $recoveryEmail, role: $role, currentPoints: $currentPoints, homeBaseId: $homeBaseId, qrToken: $qrToken, bio: $bio, photoUrl: $photoUrl, bannerUrl: $bannerUrl, following: $following, realNameVisibility: $realNameVisibility, onlineStatus: $onlineStatus, currentCheckInHallId: $currentCheckInHallId, blockedUsers: $blockedUsers, customCategories: $customCategories, lastViewedPhotoApprovals: $lastViewedPhotoApprovals, fcmTokens: $fcmTokens)';
 }
 
 
@@ -52,7 +53,7 @@ abstract mixin class $UserModelCopyWith<$Res>  {
   factory $UserModelCopyWith(UserModel value, $Res Function(UserModel) _then) = _$UserModelCopyWithImpl;
 @useResult
 $Res call({
- String uid, String email, String firstName, String lastName, String username, DateTime birthday, String? phoneNumber, String? recoveryEmail, String role, int currentPoints, String? homeBaseId, String? qrToken, String? bio, String? photoUrl, String? bannerUrl, List<String> following, String realNameVisibility, String onlineStatus, String? currentCheckInHallId, List<String> blockedUsers, List<String> customCategories,@NullableTimestampConverter() DateTime? lastViewedPhotoApprovals
+ String uid, String email, String firstName, String lastName, String username, DateTime birthday, String? phoneNumber, String? recoveryEmail, String role, int currentPoints, String? homeBaseId, String? qrToken, String? bio, String? photoUrl, String? bannerUrl, List<String> following, String realNameVisibility, String onlineStatus, String? currentCheckInHallId, List<String> blockedUsers, List<String> customCategories,@NullableTimestampConverter() DateTime? lastViewedPhotoApprovals, List<String> fcmTokens
 });
 
 
@@ -69,7 +70,7 @@ class _$UserModelCopyWithImpl<$Res>
 
 /// Create a copy of UserModel
 /// with the given fields replaced by the non-null parameter values.
-@pragma('vm:prefer-inline') @override $Res call({Object? uid = null,Object? email = null,Object? firstName = null,Object? lastName = null,Object? username = null,Object? birthday = null,Object? phoneNumber = freezed,Object? recoveryEmail = freezed,Object? role = null,Object? currentPoints = null,Object? homeBaseId = freezed,Object? qrToken = freezed,Object? bio = freezed,Object? photoUrl = freezed,Object? bannerUrl = freezed,Object? following = null,Object? realNameVisibility = null,Object? onlineStatus = null,Object? currentCheckInHallId = freezed,Object? blockedUsers = null,Object? customCategories = null,Object? lastViewedPhotoApprovals = freezed,}) {
+@pragma('vm:prefer-inline') @override $Res call({Object? uid = null,Object? email = null,Object? firstName = null,Object? lastName = null,Object? username = null,Object? birthday = null,Object? phoneNumber = freezed,Object? recoveryEmail = freezed,Object? role = null,Object? currentPoints = null,Object? homeBaseId = freezed,Object? qrToken = freezed,Object? bio = freezed,Object? photoUrl = freezed,Object? bannerUrl = freezed,Object? following = null,Object? realNameVisibility = null,Object? onlineStatus = null,Object? currentCheckInHallId = freezed,Object? blockedUsers = null,Object? customCategories = null,Object? lastViewedPhotoApprovals = freezed,Object? fcmTokens = null,}) {
   return _then(_self.copyWith(
 uid: null == uid ? _self.uid : uid // ignore: cast_nullable_to_non_nullable
 as String,email: null == email ? _self.email : email // ignore: cast_nullable_to_non_nullable
@@ -93,7 +94,8 @@ as String,currentCheckInHallId: freezed == currentCheckInHallId ? _self.currentC
 as String?,blockedUsers: null == blockedUsers ? _self.blockedUsers : blockedUsers // ignore: cast_nullable_to_non_nullable
 as List<String>,customCategories: null == customCategories ? _self.customCategories : customCategories // ignore: cast_nullable_to_non_nullable
 as List<String>,lastViewedPhotoApprovals: freezed == lastViewedPhotoApprovals ? _self.lastViewedPhotoApprovals : lastViewedPhotoApprovals // ignore: cast_nullable_to_non_nullable
-as DateTime?,
+as DateTime?,fcmTokens: null == fcmTokens ? _self.fcmTokens : fcmTokens // ignore: cast_nullable_to_non_nullable
+as List<String>,
   ));
 }
 
@@ -178,10 +180,10 @@ return $default(_that);case _:
 /// }
 /// ```
 
-@optionalTypeArgs TResult maybeWhen<TResult extends Object?>(TResult Function( String uid,  String email,  String firstName,  String lastName,  String username,  DateTime birthday,  String? phoneNumber,  String? recoveryEmail,  String role,  int currentPoints,  String? homeBaseId,  String? qrToken,  String? bio,  String? photoUrl,  String? bannerUrl,  List<String> following,  String realNameVisibility,  String onlineStatus,  String? currentCheckInHallId,  List<String> blockedUsers,  List<String> customCategories, @NullableTimestampConverter()  DateTime? lastViewedPhotoApprovals)?  $default,{required TResult orElse(),}) {final _that = this;
+@optionalTypeArgs TResult maybeWhen<TResult extends Object?>(TResult Function( String uid,  String email,  String firstName,  String lastName,  String username,  DateTime birthday,  String? phoneNumber,  String? recoveryEmail,  String role,  int currentPoints,  String? homeBaseId,  String? qrToken,  String? bio,  String? photoUrl,  String? bannerUrl,  List<String> following,  String realNameVisibility,  String onlineStatus,  String? currentCheckInHallId,  List<String> blockedUsers,  List<String> customCategories, @NullableTimestampConverter()  DateTime? lastViewedPhotoApprovals,  List<String> fcmTokens)?  $default,{required TResult orElse(),}) {final _that = this;
 switch (_that) {
 case _UserModel() when $default != null:
-return $default(_that.uid,_that.email,_that.firstName,_that.lastName,_that.username,_that.birthday,_that.phoneNumber,_that.recoveryEmail,_that.role,_that.currentPoints,_that.homeBaseId,_that.qrToken,_that.bio,_that.photoUrl,_that.bannerUrl,_that.following,_that.realNameVisibility,_that.onlineStatus,_that.currentCheckInHallId,_that.blockedUsers,_that.customCategories,_that.lastViewedPhotoApprovals);case _:
+return $default(_that.uid,_that.email,_that.firstName,_that.lastName,_that.username,_that.birthday,_that.phoneNumber,_that.recoveryEmail,_that.role,_that.currentPoints,_that.homeBaseId,_that.qrToken,_that.bio,_that.photoUrl,_that.bannerUrl,_that.following,_that.realNameVisibility,_that.onlineStatus,_that.currentCheckInHallId,_that.blockedUsers,_that.customCategories,_that.lastViewedPhotoApprovals,_that.fcmTokens);case _:
   return orElse();
 
 }
@@ -199,10 +201,10 @@ return $default(_that.uid,_that.email,_that.firstName,_that.lastName,_that.usern
 /// }
 /// ```
 
-@optionalTypeArgs TResult when<TResult extends Object?>(TResult Function( String uid,  String email,  String firstName,  String lastName,  String username,  DateTime birthday,  String? phoneNumber,  String? recoveryEmail,  String role,  int currentPoints,  String? homeBaseId,  String? qrToken,  String? bio,  String? photoUrl,  String? bannerUrl,  List<String> following,  String realNameVisibility,  String onlineStatus,  String? currentCheckInHallId,  List<String> blockedUsers,  List<String> customCategories, @NullableTimestampConverter()  DateTime? lastViewedPhotoApprovals)  $default,) {final _that = this;
+@optionalTypeArgs TResult when<TResult extends Object?>(TResult Function( String uid,  String email,  String firstName,  String lastName,  String username,  DateTime birthday,  String? phoneNumber,  String? recoveryEmail,  String role,  int currentPoints,  String? homeBaseId,  String? qrToken,  String? bio,  String? photoUrl,  String? bannerUrl,  List<String> following,  String realNameVisibility,  String onlineStatus,  String? currentCheckInHallId,  List<String> blockedUsers,  List<String> customCategories, @NullableTimestampConverter()  DateTime? lastViewedPhotoApprovals,  List<String> fcmTokens)  $default,) {final _that = this;
 switch (_that) {
 case _UserModel():
-return $default(_that.uid,_that.email,_that.firstName,_that.lastName,_that.username,_that.birthday,_that.phoneNumber,_that.recoveryEmail,_that.role,_that.currentPoints,_that.homeBaseId,_that.qrToken,_that.bio,_that.photoUrl,_that.bannerUrl,_that.following,_that.realNameVisibility,_that.onlineStatus,_that.currentCheckInHallId,_that.blockedUsers,_that.customCategories,_that.lastViewedPhotoApprovals);case _:
+return $default(_that.uid,_that.email,_that.firstName,_that.lastName,_that.username,_that.birthday,_that.phoneNumber,_that.recoveryEmail,_that.role,_that.currentPoints,_that.homeBaseId,_that.qrToken,_that.bio,_that.photoUrl,_that.bannerUrl,_that.following,_that.realNameVisibility,_that.onlineStatus,_that.currentCheckInHallId,_that.blockedUsers,_that.customCategories,_that.lastViewedPhotoApprovals,_that.fcmTokens);case _:
   throw StateError('Unexpected subclass');
 
 }
@@ -219,10 +221,10 @@ return $default(_that.uid,_that.email,_that.firstName,_that.lastName,_that.usern
 /// }
 /// ```
 
-@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>(TResult? Function( String uid,  String email,  String firstName,  String lastName,  String username,  DateTime birthday,  String? phoneNumber,  String? recoveryEmail,  String role,  int currentPoints,  String? homeBaseId,  String? qrToken,  String? bio,  String? photoUrl,  String? bannerUrl,  List<String> following,  String realNameVisibility,  String onlineStatus,  String? currentCheckInHallId,  List<String> blockedUsers,  List<String> customCategories, @NullableTimestampConverter()  DateTime? lastViewedPhotoApprovals)?  $default,) {final _that = this;
+@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>(TResult? Function( String uid,  String email,  String firstName,  String lastName,  String username,  DateTime birthday,  String? phoneNumber,  String? recoveryEmail,  String role,  int currentPoints,  String? homeBaseId,  String? qrToken,  String? bio,  String? photoUrl,  String? bannerUrl,  List<String> following,  String realNameVisibility,  String onlineStatus,  String? currentCheckInHallId,  List<String> blockedUsers,  List<String> customCategories, @NullableTimestampConverter()  DateTime? lastViewedPhotoApprovals,  List<String> fcmTokens)?  $default,) {final _that = this;
 switch (_that) {
 case _UserModel() when $default != null:
-return $default(_that.uid,_that.email,_that.firstName,_that.lastName,_that.username,_that.birthday,_that.phoneNumber,_that.recoveryEmail,_that.role,_that.currentPoints,_that.homeBaseId,_that.qrToken,_that.bio,_that.photoUrl,_that.bannerUrl,_that.following,_that.realNameVisibility,_that.onlineStatus,_that.currentCheckInHallId,_that.blockedUsers,_that.customCategories,_that.lastViewedPhotoApprovals);case _:
+return $default(_that.uid,_that.email,_that.firstName,_that.lastName,_that.username,_that.birthday,_that.phoneNumber,_that.recoveryEmail,_that.role,_that.currentPoints,_that.homeBaseId,_that.qrToken,_that.bio,_that.photoUrl,_that.bannerUrl,_that.following,_that.realNameVisibility,_that.onlineStatus,_that.currentCheckInHallId,_that.blockedUsers,_that.customCategories,_that.lastViewedPhotoApprovals,_that.fcmTokens);case _:
   return null;
 
 }
@@ -234,7 +236,7 @@ return $default(_that.uid,_that.email,_that.firstName,_that.lastName,_that.usern
 @JsonSerializable()
 
 class _UserModel implements UserModel {
-  const _UserModel({required this.uid, required this.email, required this.firstName, required this.lastName, required this.username, required this.birthday, this.phoneNumber, this.recoveryEmail, this.role = 'player', this.currentPoints = 0, this.homeBaseId, this.qrToken, this.bio, this.photoUrl, this.bannerUrl, final  List<String> following = const [], this.realNameVisibility = 'Private', this.onlineStatus = 'Online', this.currentCheckInHallId, final  List<String> blockedUsers = const [], final  List<String> customCategories = const [], @NullableTimestampConverter() this.lastViewedPhotoApprovals}): _following = following,_blockedUsers = blockedUsers,_customCategories = customCategories;
+  const _UserModel({required this.uid, required this.email, required this.firstName, required this.lastName, required this.username, required this.birthday, this.phoneNumber, this.recoveryEmail, this.role = 'player', this.currentPoints = 0, this.homeBaseId, this.qrToken, this.bio, this.photoUrl, this.bannerUrl, final  List<String> following = const [], this.realNameVisibility = 'Private', this.onlineStatus = 'Online', this.currentCheckInHallId, final  List<String> blockedUsers = const [], final  List<String> customCategories = const [], @NullableTimestampConverter() this.lastViewedPhotoApprovals, final  List<String> fcmTokens = const []}): _following = following,_blockedUsers = blockedUsers,_customCategories = customCategories,_fcmTokens = fcmTokens;
   factory _UserModel.fromJson(Map<String, dynamic> json) => _$UserModelFromJson(json);
 
 @override final  String uid;
@@ -282,6 +284,15 @@ class _UserModel implements UserModel {
 
 // Array of Custom Tags/Categories the user has saved
 @override@NullableTimestampConverter() final  DateTime? lastViewedPhotoApprovals;
+// Track when the worker last viewed pending photos
+ final  List<String> _fcmTokens;
+// Track when the worker last viewed pending photos
+@override@JsonKey() List<String> get fcmTokens {
+  if (_fcmTokens is EqualUnmodifiableListView) return _fcmTokens;
+  // ignore: implicit_dynamic_type
+  return EqualUnmodifiableListView(_fcmTokens);
+}
+
 
 /// Create a copy of UserModel
 /// with the given fields replaced by the non-null parameter values.
@@ -296,16 +307,16 @@ Map<String, dynamic> toJson() {
 
 @override
 bool operator ==(Object other) {
-  return identical(this, other) || (other.runtimeType == runtimeType&&other is _UserModel&&(identical(other.uid, uid) || other.uid == uid)&&(identical(other.email, email) || other.email == email)&&(identical(other.firstName, firstName) || other.firstName == firstName)&&(identical(other.lastName, lastName) || other.lastName == lastName)&&(identical(other.username, username) || other.username == username)&&(identical(other.birthday, birthday) || other.birthday == birthday)&&(identical(other.phoneNumber, phoneNumber) || other.phoneNumber == phoneNumber)&&(identical(other.recoveryEmail, recoveryEmail) || other.recoveryEmail == recoveryEmail)&&(identical(other.role, role) || other.role == role)&&(identical(other.currentPoints, currentPoints) || other.currentPoints == currentPoints)&&(identical(other.homeBaseId, homeBaseId) || other.homeBaseId == homeBaseId)&&(identical(other.qrToken, qrToken) || other.qrToken == qrToken)&&(identical(other.bio, bio) || other.bio == bio)&&(identical(other.photoUrl, photoUrl) || other.photoUrl == photoUrl)&&(identical(other.bannerUrl, bannerUrl) || other.bannerUrl == bannerUrl)&&const DeepCollectionEquality().equals(other._following, _following)&&(identical(other.realNameVisibility, realNameVisibility) || other.realNameVisibility == realNameVisibility)&&(identical(other.onlineStatus, onlineStatus) || other.onlineStatus == onlineStatus)&&(identical(other.currentCheckInHallId, currentCheckInHallId) || other.currentCheckInHallId == currentCheckInHallId)&&const DeepCollectionEquality().equals(other._blockedUsers, _blockedUsers)&&const DeepCollectionEquality().equals(other._customCategories, _customCategories)&&(identical(other.lastViewedPhotoApprovals, lastViewedPhotoApprovals) || other.lastViewedPhotoApprovals == lastViewedPhotoApprovals));
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is _UserModel&&(identical(other.uid, uid) || other.uid == uid)&&(identical(other.email, email) || other.email == email)&&(identical(other.firstName, firstName) || other.firstName == firstName)&&(identical(other.lastName, lastName) || other.lastName == lastName)&&(identical(other.username, username) || other.username == username)&&(identical(other.birthday, birthday) || other.birthday == birthday)&&(identical(other.phoneNumber, phoneNumber) || other.phoneNumber == phoneNumber)&&(identical(other.recoveryEmail, recoveryEmail) || other.recoveryEmail == recoveryEmail)&&(identical(other.role, role) || other.role == role)&&(identical(other.currentPoints, currentPoints) || other.currentPoints == currentPoints)&&(identical(other.homeBaseId, homeBaseId) || other.homeBaseId == homeBaseId)&&(identical(other.qrToken, qrToken) || other.qrToken == qrToken)&&(identical(other.bio, bio) || other.bio == bio)&&(identical(other.photoUrl, photoUrl) || other.photoUrl == photoUrl)&&(identical(other.bannerUrl, bannerUrl) || other.bannerUrl == bannerUrl)&&const DeepCollectionEquality().equals(other._following, _following)&&(identical(other.realNameVisibility, realNameVisibility) || other.realNameVisibility == realNameVisibility)&&(identical(other.onlineStatus, onlineStatus) || other.onlineStatus == onlineStatus)&&(identical(other.currentCheckInHallId, currentCheckInHallId) || other.currentCheckInHallId == currentCheckInHallId)&&const DeepCollectionEquality().equals(other._blockedUsers, _blockedUsers)&&const DeepCollectionEquality().equals(other._customCategories, _customCategories)&&(identical(other.lastViewedPhotoApprovals, lastViewedPhotoApprovals) || other.lastViewedPhotoApprovals == lastViewedPhotoApprovals)&&const DeepCollectionEquality().equals(other._fcmTokens, _fcmTokens));
 }
 
 @JsonKey(includeFromJson: false, includeToJson: false)
 @override
-int get hashCode => Object.hashAll([runtimeType,uid,email,firstName,lastName,username,birthday,phoneNumber,recoveryEmail,role,currentPoints,homeBaseId,qrToken,bio,photoUrl,bannerUrl,const DeepCollectionEquality().hash(_following),realNameVisibility,onlineStatus,currentCheckInHallId,const DeepCollectionEquality().hash(_blockedUsers),const DeepCollectionEquality().hash(_customCategories),lastViewedPhotoApprovals]);
+int get hashCode => Object.hashAll([runtimeType,uid,email,firstName,lastName,username,birthday,phoneNumber,recoveryEmail,role,currentPoints,homeBaseId,qrToken,bio,photoUrl,bannerUrl,const DeepCollectionEquality().hash(_following),realNameVisibility,onlineStatus,currentCheckInHallId,const DeepCollectionEquality().hash(_blockedUsers),const DeepCollectionEquality().hash(_customCategories),lastViewedPhotoApprovals,const DeepCollectionEquality().hash(_fcmTokens)]);
 
 @override
 String toString() {
-  return 'UserModel(uid: $uid, email: $email, firstName: $firstName, lastName: $lastName, username: $username, birthday: $birthday, phoneNumber: $phoneNumber, recoveryEmail: $recoveryEmail, role: $role, currentPoints: $currentPoints, homeBaseId: $homeBaseId, qrToken: $qrToken, bio: $bio, photoUrl: $photoUrl, bannerUrl: $bannerUrl, following: $following, realNameVisibility: $realNameVisibility, onlineStatus: $onlineStatus, currentCheckInHallId: $currentCheckInHallId, blockedUsers: $blockedUsers, customCategories: $customCategories, lastViewedPhotoApprovals: $lastViewedPhotoApprovals)';
+  return 'UserModel(uid: $uid, email: $email, firstName: $firstName, lastName: $lastName, username: $username, birthday: $birthday, phoneNumber: $phoneNumber, recoveryEmail: $recoveryEmail, role: $role, currentPoints: $currentPoints, homeBaseId: $homeBaseId, qrToken: $qrToken, bio: $bio, photoUrl: $photoUrl, bannerUrl: $bannerUrl, following: $following, realNameVisibility: $realNameVisibility, onlineStatus: $onlineStatus, currentCheckInHallId: $currentCheckInHallId, blockedUsers: $blockedUsers, customCategories: $customCategories, lastViewedPhotoApprovals: $lastViewedPhotoApprovals, fcmTokens: $fcmTokens)';
 }
 
 
@@ -316,7 +327,7 @@ abstract mixin class _$UserModelCopyWith<$Res> implements $UserModelCopyWith<$Re
   factory _$UserModelCopyWith(_UserModel value, $Res Function(_UserModel) _then) = __$UserModelCopyWithImpl;
 @override @useResult
 $Res call({
- String uid, String email, String firstName, String lastName, String username, DateTime birthday, String? phoneNumber, String? recoveryEmail, String role, int currentPoints, String? homeBaseId, String? qrToken, String? bio, String? photoUrl, String? bannerUrl, List<String> following, String realNameVisibility, String onlineStatus, String? currentCheckInHallId, List<String> blockedUsers, List<String> customCategories,@NullableTimestampConverter() DateTime? lastViewedPhotoApprovals
+ String uid, String email, String firstName, String lastName, String username, DateTime birthday, String? phoneNumber, String? recoveryEmail, String role, int currentPoints, String? homeBaseId, String? qrToken, String? bio, String? photoUrl, String? bannerUrl, List<String> following, String realNameVisibility, String onlineStatus, String? currentCheckInHallId, List<String> blockedUsers, List<String> customCategories,@NullableTimestampConverter() DateTime? lastViewedPhotoApprovals, List<String> fcmTokens
 });
 
 
@@ -333,7 +344,7 @@ class __$UserModelCopyWithImpl<$Res>
 
 /// Create a copy of UserModel
 /// with the given fields replaced by the non-null parameter values.
-@override @pragma('vm:prefer-inline') $Res call({Object? uid = null,Object? email = null,Object? firstName = null,Object? lastName = null,Object? username = null,Object? birthday = null,Object? phoneNumber = freezed,Object? recoveryEmail = freezed,Object? role = null,Object? currentPoints = null,Object? homeBaseId = freezed,Object? qrToken = freezed,Object? bio = freezed,Object? photoUrl = freezed,Object? bannerUrl = freezed,Object? following = null,Object? realNameVisibility = null,Object? onlineStatus = null,Object? currentCheckInHallId = freezed,Object? blockedUsers = null,Object? customCategories = null,Object? lastViewedPhotoApprovals = freezed,}) {
+@override @pragma('vm:prefer-inline') $Res call({Object? uid = null,Object? email = null,Object? firstName = null,Object? lastName = null,Object? username = null,Object? birthday = null,Object? phoneNumber = freezed,Object? recoveryEmail = freezed,Object? role = null,Object? currentPoints = null,Object? homeBaseId = freezed,Object? qrToken = freezed,Object? bio = freezed,Object? photoUrl = freezed,Object? bannerUrl = freezed,Object? following = null,Object? realNameVisibility = null,Object? onlineStatus = null,Object? currentCheckInHallId = freezed,Object? blockedUsers = null,Object? customCategories = null,Object? lastViewedPhotoApprovals = freezed,Object? fcmTokens = null,}) {
   return _then(_UserModel(
 uid: null == uid ? _self.uid : uid // ignore: cast_nullable_to_non_nullable
 as String,email: null == email ? _self.email : email // ignore: cast_nullable_to_non_nullable
@@ -357,7 +368,8 @@ as String,currentCheckInHallId: freezed == currentCheckInHallId ? _self.currentC
 as String?,blockedUsers: null == blockedUsers ? _self._blockedUsers : blockedUsers // ignore: cast_nullable_to_non_nullable
 as List<String>,customCategories: null == customCategories ? _self._customCategories : customCategories // ignore: cast_nullable_to_non_nullable
 as List<String>,lastViewedPhotoApprovals: freezed == lastViewedPhotoApprovals ? _self.lastViewedPhotoApprovals : lastViewedPhotoApprovals // ignore: cast_nullable_to_non_nullable
-as DateTime?,
+as DateTime?,fcmTokens: null == fcmTokens ? _self._fcmTokens : fcmTokens // ignore: cast_nullable_to_non_nullable
+as List<String>,
   ));
 }